+++ dailyreportsystem/dailyreportsystem/urls.py
@@ -35,7 +35,4 @@ urlpatterns = [
     path('', RedirectView.as_view(url='drs/')),
     path('login/', views.loginUser, name='login-user'),
     path('logout/', views.logoutUser, name='logout-user'),
-] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
-
-if settings.DEBUG:
-    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
