+++ dailyreportsystem/drs/templates/profile/profile.html
@@ -27,7 +27,7 @@
                     <!-- END SIDEBAR USER TITLE -->
                     <!-- SIDEBAR BUTTONS -->
                     <div class="profile-userbuttons">
-                        <a href="/drs/profile_update/"  class="btn btn-primary" >Update Profile User</a>
+                        <a href="{% url 'profile_update'%}  class="btn btn-primary" >Update Profile User</a>
                     </div>
                     <!-- END SIDEBAR BUTTONS -->
                     <!-- SIDEBAR MENU -->
