--- dailyreportsystem/drs/templates/sign_up/inform_to_verify.html
@@ -1,2 +1,0 @@
-{% load i18n %}
-<h2>{% trans 'inform to verify' %}</h2>
