--- dailyreportsystem/drs/templates/sign_up/register.html
@@ -1,60 +1,0 @@
-{% extends "base_generic.html" %}
-{% load static %}
-{% block title %}<title>Sign up</title>{% endblock %}
-
-{% block content %}
-
-<div class="container">
-	<div class="main">
-		<div class="registration">
-			<div class="registration_left">
-			<h2>new user? <span> create an account </span></h2>
-			 <div class="registration_form">
-			 <!-- Form -->
-				<form id="registration_left" method="post">
-
-					{% if form.errors %}
-						{% for field in form %}
-							{% for error in field.errors %}
-								<div class="alert alert-danger">
-									<p><strong>{{ error|escape }}</strong></p>
-								</div>
-							{% endfor %}
-						{% endfor %}
-						{% for error in form.non_field_errors %}
-							<div class="alert alert-danger">
-								<p><strong>{{ error|escape }}</strong></p>
-							</div>
-						{% endfor %}
-					{% endif %}
-
-					{% csrf_token %}
-					<div>
-						<label>
-							<input placeholder="email address:" type="email" name="email" tabindex="3" required>
-						</label>
-					</div>
-					<div>
-						<label>
-							<input placeholder="password" type="password" name="password1" tabindex="4" required>
-						</label>
-					</div>
-					<div>
-						<label>
-							<input placeholder="retype password" type="password" name="password2" tabindex="4" required>
-						</label>
-					</div>
-					<div>
-						<input type="submit" value="create an account" id="register-submit">
-					</div>
-				</form>
-			</div>
-		</div>
-		<div class="clearfix"></div>
-		</div>
-	</div>
-</div>
-
-<script src="{% static '/js/register.js' %}"></script>
-
-{% endblock %}
