--- dailyreportsystem/drs/templates/sign_up/verification_fail.html
@@ -1,2 +1,0 @@
-{% load i18n %}
-<h2>{% trans 'fail_activation' %}</h2>
