--- dailyreportsystem/drs/templates/sign_up/verification_success.html
@@ -1,2 +1,0 @@
-{% load i18n %}
-<h2>{% trans 'success_activation' %}</h2>
