+++ dailyreportsystem/templates/sign_up/inform_to_verify.html
@@ -1,0 +1,17 @@
+{% load i18n %}
+<head>
+    <title>Bootstrap Example</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+</head>
+<body>
+<div class="container">
+    <h2>Alert User To Check Email</h2>
+    <div class="alert alert-info">
+        <strong> PLEASEEEEEEE!!!!!!</strong> You should <a href="https://gmail.com/" class="alert-link">check your email</a>.
+    </div>
+</div>
+</body>
