+++ dailyreportsystem/templates/sign_up/password_reset_done.html
@@ -1,0 +1,5 @@
+{% extends "base_generic.html" %}
+
+{% block content %}
+  <p>We've emailed you instructions for setting your password. If they haven't arrived in a few minutes, check your spam folder.</p>
+{% endblock %}
