+++ dailyreportsystem/templates/sign_up/register.html
@@ -1,0 +1,63 @@
+{% extends "base_generic.html" %}
+{% load i18n %}
+{% load static %}
+{% block media %}
+    {% block title %}<title>Register User</title>{% endblock %}
+{% endblock %}
+{% block content %}
+<div class="container">
+	<div class="row" style="margin-top: 50px;">
+        <div class="col-3"></div>
+        <div class="col-5">
+            <!-- Material form login -->
+            <div class="card">
+                <h5 class="card-header info-color white-text text-center py-4">
+                    <strong>Sign in</strong>
+                </h5>
+                <!--Card content-->
+                <div class="card-body px-lg-5 pt-0">
+                    <!-- Form -->
+                    <form class="text-center" style="color: #757575;" method="post">
+                        {% if form.errors %}
+                            {% for field in form %}
+                                {% for error in field.errors %}
+                                    <div class="alert alert-danger">
+                                        <p><strong>{{ error|escape }}</strong></p>
+                                    </div>
+                                {% endfor %}
+                            {% endfor %}
+                            {% for error in form.non_field_errors %}
+                                <div class="alert alert-danger">
+                                    <p><strong>{{ error|escape }}</strong></p>
+                                </div>
+                            {% endfor %}
+                        {% endif %}
+                        {% csrf_token %}
+                        <!-- Email -->
+                        <div class="md-form">
+                            <input type="email" id="materialLoginFormEmail" class="form-control" name="email" required>
+                            <label for="materialLoginFormEmail">E-mail</label>
+                        </div>
+                        <!-- Password -->
+                        <div class="md-form">
+                            <input type="password" id="materialLoginFormPassword" class="form-control" name="password1" required>
+                            <label for="materialLoginFormPassword">Password</label>
+                        </div>
+                        <div class="md-form">
+                            <input type="password" id="materialLoginFormPassword" class="form-control" name="password2" required>
+                            <label for="materialLoginFormPassword">Retype Password</label>
+                        </div>
+                        <!-- Sign in button -->
+                        <button class="btn btn-outline-info btn-rounded btn-block my-4 waves-effect z-depth-0" type="submit">Create an account</button>
+                    </form>
+                    <!-- Form -->
+                </div>
+            </div>
+            <!-- Material form login -->
+        </div>
+    </div>
+</div>
+{% block js%}
+    <script src="{% static '/js/register.js' %}"></script>
+{% endblock %}
+{% endblock %}
